refactor: フォルダ名をcontextsからprovidersに変更
