+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import Layout from '../components/Layout';
 import TodoList from '../components/TodoList';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../providers/AuthProvider';
 import type { NextPage } from 'next';
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
