+++ src/pages/user.tsx
@@ -1,6 +1,6 @@
 import AccountSettings from '../components/AccountSettings';
 import Layout from '../components/Layout';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../providers/AuthProvider';
 import type { NextPage } from 'next';
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
